Major app-reorganization, it all works again.
