+++ .vscode/launch.json
@@ -10,7 +10,6 @@
             "request": "launch",
             "module": "flask",
             "env": {
-                "FLASK_APP": "main.py",
                 "FLASK_DEBUG": "1"
             },
             "args": [
