+++ config.py
@@ -1,8 +1,8 @@
 from os import environ, path
-# from dotenv import load_dotenv
+from dotenv import load_dotenv
 
 basedir = path.abspath(path.dirname(__file__))
-# load_dotenv(path.join(basedir, '.env'))
+load_dotenv(path.join(basedir, '.env'))
 
 
 class BaseConfig(object):
@@ -34,7 +34,7 @@ class DevConfig(BaseConfig):
     FLASK_DEBUG = True
     TESTING = True
     DEBUG = True
-    SQLALCHEMY_DATABASE_URI = 'sqlite:///wallet.sqlite'
+    SQLALCHEMY_DATABASE_URI = 'sqlite:///./wallet.sqlite'
 
 
 class TestConfig(BaseConfig):
