+++ tests/test_character.py
@@ -1,34 +1,39 @@
 """ Tests related to CRUD operations for Characters """
 import pytest
-
-from main import app, db, Character
+from app import db, create_app
+from app.character.models import Character
 from config import TestConfig
 
 
-@pytest.fixture
-def client(scope='function'):
+@pytest.fixture(scope='session')
+def app():
+    app = create_app()
     config = TestConfig()
     app.config.from_object(config)
+    return app
 
-    with app.test_client() as client:
-        with app.app_context():
-            db.create_all()
+
+@pytest.fixture(scope='function')
+def client(app):
+    with app.app_context():
+        client = app.test_client()
+        db.create_all()
+
         yield client
         db.drop_all()
 
 
-@pytest.fixture
-def client_loaded():
-    config = TestConfig()
-    app.config.from_object(config)
+@pytest.fixture(scope='function')
+def client_loaded(app):
+    with app.app_context():
+        client_loaded = app.test_client()
+        db.create_all()
 
-    with app.test_client() as client:
-        with app.app_context():
-            db.create_all()
         db.session.add(Character(id=1, name='Paladin', is_dead=False))
         db.session.add(Character(id=2, name='Rogue', is_dead=False))
         db.session.add(Character(id=3, name='Fighter', is_dead=False))
-        yield client
+
+        yield client_loaded
         db.drop_all()
